4.3.2: bound Retry-After and the poll's sleep, refuse a timeout where set
Measured on 4.3.1 against a local origin (2026-09-23):
- Retry-After 9223372036854775807 overflowed delayFor's milliseconds into a
  float, and the int return type failed the call with a raw TypeError;
  9223372036854775 scheduled a wait of millennia. Past 2**31 - 1 ms, the
  .NET bound, a Retry-After now waits the backoff, still a rate limit.
- A NaN, infinite, sub-millisecond or 1e300 timeout was accepted by Options
  and per call, then Guzzle refused it on every request. Refused where it is
  set now, as is anything from 9e15 s, whose milliseconds overflow an int.
- pollDeviceToken slept the full interval: sleep() takes a C unsigned int,
  so 2147483647 held the poll 68 years and 4294967297 wrapped to 1 s. It
  now sleeps min(interval, time left), never negative, through
  time_nanosleep in parts (UMAN-4483), with the two corpus cases.

Also ports internetdata's test that leaves out ONE required member per case.
Twelve mutants, eleven red; the survivor, `$m['interval'] ?? 5`, is dead
code behind the required-member table, whose own mutant is red.

+++ tests/OauthStub.php
@@ -34,12 +34,12 @@ final class OauthStub
      */
     public array $requests = [];
 
-    /** @var list<int> Every wait the poll took, in seconds. */
+    /** @var list<float> Every wait the poll took, in seconds. */
     public array $waits = [];
 
     public readonly GuzzleClient $client;
 
-    private int $elapsed = 0;
+    private float $elapsed = 0.0;
     private int $reads = 0;
 
     /** @param list<array<string, mixed>> $replies Each `status` plus `body` (sent as JSON) or `rawBody`. */
@@ -55,9 +55,9 @@ final class OauthStub
             if (++$this->reads > 2 * self::BOUND) {
                 self::trip("read the clock {$this->reads} times");
             }
-            return (float) $this->elapsed;
+            return $this->elapsed;
         };
-        $sleep = function (int $seconds): void {
+        $sleep = function (float $seconds): void {
             if (count($this->waits) === self::BOUND) {
                 self::trip('waited more than ' . self::BOUND . ' times');
             }
